% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necofs-package.R
\docType{package}
\name{cascophys-package}
\alias{cascophys-package}
\alias{casco}
\alias{cascophys}
\title{cascophys-package}
\description{
Convenience tools for accessing Casco Bay centric FVCOM NetCDF files
}
