% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NE_Physics.R
\name{NECOFSPhysics}
\alias{NECOFSPhysics}
\title{Instantiate a NE_Physics R6 object}
\usage{
NECOFSPhysics(
  filename = file.path("http://www.smast.umassd.edu:8080",
    "thredds/dodsC/models/fvcom/NECOFS/Archive/NECOFS_GOM/2019", "gom4_201901.nc"),
  ...
)
}
\arguments{
\item{filename}{character either a filename or OpenDAP URL}

\item{...}{other arguments passed to \code{NE_Physics$new(filename, ...)}}
}
\value{
NE_Physics R6 Reference Object
}
\description{
Instantiate a NE_Physics R6 object
}
