% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NE_Physics.R
\name{random_point}
\alias{random_point}
\title{Generate  listing of one or more random points}
\usage{
random_point(X, n = 1, select_time = c("first", "last", "random")[1])
}
\arguments{
\item{X}{NE_Physics object}

\item{n}{numeric the number of random points to generate}

\item{select_time}{character, one of 'first' (default), 'last' or 'random'}
}
\value{
sf POINT XYZ object with \code{n} features
}
\description{
Generate  listing of one or more random points
}
